fix(tests) dual K8s cluster test support
