+++ test/framework/k8s_clusters.go
@@ -1,0 +1,108 @@
+package framework
+
+import (
+	"fmt"
+	"github.com/gruntwork-io/terratest/modules/k8s"
+	"github.com/gruntwork-io/terratest/modules/testing"
+	"os"
+)
+
+type K8sClusters struct {
+	t        testing.TestingT
+	clusters map[string]*K8sCluster
+	verbose  bool
+}
+
+// NewK8sTest gets the number of the clusters to use in the tests, and the pattern
+// to locate the KUBECONFIG for them. The second argument can be empty
+func NewK8sClusters(clusterNames []string, verbose bool) (Clusters, error) {
+	if len(clusterNames) < 1 || len(clusterNames) > maxClusters {
+		return nil, fmt.Errorf("Invalid cluster number. Should be in the range [1,3], but it is %d", len(clusterNames))
+	}
+
+	t := NewTestingT()
+	clusters := map[string]*K8sCluster{}
+	for _, name := range clusterNames {
+		options, err := NewKumactlOptions(t, name, "", "", verbose)
+		if err != nil {
+			return nil, err
+		}
+		clusters[name] = &K8sCluster{
+			t:                   t,
+			name:                name,
+			kubeconfig:          os.ExpandEnv(fmt.Sprintf(defaultKubeConfigPathPattern, name)),
+			kumactl:             options,
+			verbose:             verbose,
+			forwardedPortsChans: map[uint32]chan struct{}{},
+		}
+	}
+
+	return &K8sClusters{
+		t:        t,
+		clusters: clusters,
+		verbose:  verbose,
+	}, nil
+}
+
+func (cs *K8sClusters) GetCluster(name string) Cluster {
+	c, found := cs.clusters[name]
+	if !found {
+		return nil
+	}
+	return c
+}
+
+func (cs *K8sClusters) DeployKuma() error {
+	for name, c := range cs.clusters {
+		if err := c.DeployKuma(); err != nil {
+			return fmt.Errorf("Deploy Kuma on %s failed: %v", name, err)
+		}
+	}
+	return nil
+}
+
+func (cs *K8sClusters) VerifyKuma() error {
+	for name, c := range cs.clusters {
+		if err := c.VerifyKuma(); err != nil {
+			return fmt.Errorf("Verify Kuma on %s failed: %v", name, err)
+		}
+	}
+	return nil
+}
+
+func (cs *K8sClusters) GetKumaCPLogs() (string, error) {
+	logs := ""
+	for name, c := range cs.clusters {
+		log, err := c.GetKumaCPLogs()
+		if err != nil {
+			return "", fmt.Errorf("Verify Kuma on %s failed: %v", name, err)
+		}
+		logs = logs + "========== " + name + " ==========\n" + log + "\n"
+	}
+	return logs, nil
+}
+
+func (cs *K8sClusters) DeleteKuma() error {
+	for name, c := range cs.clusters {
+		if err := c.DeleteKuma(); err != nil {
+			return fmt.Errorf("Delete Kuma on %s failed: %v", name, err)
+		}
+	}
+	return nil
+}
+
+func (cs *K8sClusters) GetKubectlOptions(namespace ...string) *k8s.KubectlOptions {
+	fmt.Println("Not supported at this level.")
+	return nil
+}
+
+func (c *K8sClusters) GetTesting() testing.TestingT {
+	return c.t
+}
+
+
+
+func IsK8sClustersStarted() bool {
+	_, found := os.LookupEnv(envK8SCLUSTERS)
+	return found
+}
